Refactor OSPF tests and add base class
