+++ manticore/tests/integration/org.opennaas.router.tests.capability.ospf/src/test/java/org/opennaas/router/tests/capability/ospf/OSPFIntegrationTest.java
@@ -3,25 +3,23 @@
  */
 package org.opennaas.router.tests.capability.ospf;
 
-import static org.openengsb.labs.paxexam.karaf.options.KarafDistributionOption.*;
-
-import static org.ops4j.pax.exam.CoreOptions.maven;
-import static org.ops4j.pax.exam.CoreOptions.mavenBundle;
+import static net.i2cat.nexus.tests.OpennaasExamOptions.includeFeatures;
+import static net.i2cat.nexus.tests.OpennaasExamOptions.noConsole;
+import static net.i2cat.nexus.tests.OpennaasExamOptions.opennaasDistributionConfiguration;
+import static org.openengsb.labs.paxexam.karaf.options.KarafDistributionOption.keepRuntimeFolder;
 import static org.ops4j.pax.exam.CoreOptions.options;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
+
 import javax.inject.Inject;
 
 import net.i2cat.nexus.tests.ResourceHelper;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.runner.RunWith;
-import org.opennaas.core.protocols.sessionmanager.impl.ProtocolSessionManager;
 import org.opennaas.core.resources.ILifecycle.State;
 import org.opennaas.core.resources.IResource;
 import org.opennaas.core.resources.IResourceManager;
@@ -35,9 +33,9 @@ import org.opennaas.core.resources.protocol.IProtocolSessionManager;
 import org.opennaas.core.resources.protocol.ProtocolException;
 import org.opennaas.core.resources.protocol.ProtocolSessionContext;
 import org.ops4j.pax.exam.Option;
-import org.ops4j.pax.exam.util.Filter;
 import org.ops4j.pax.exam.junit.Configuration;
 import org.ops4j.pax.exam.junit.JUnit4TestRunner;
+import org.ops4j.pax.exam.util.Filter;
 import org.osgi.framework.BundleContext;
 import org.osgi.service.blueprint.container.BlueprintContainer;
 
@@ -47,53 +45,40 @@ import org.osgi.service.blueprint.container.BlueprintContainer;
 @RunWith(JUnit4TestRunner.class)
 public abstract class OSPFIntegrationTest
 {
-	protected static final String			ACTION_NAME				= "junos";
-	protected static final String			CAPABILITY_URI			= "mock://user:pass@host.net:2212/mocksubsystem";
-	protected static final String			QUEUE_CAPABILIY_TYPE	= "queue";
-	protected static final String			OSPF_CAPABILIY_TYPE		= "ospf";
-	protected static final String			OSPF_CAPABILIY_VERSION	= "10.10";
-	protected static final String			RESOURCE_TYPE			= "router";
-	protected static final String			RESOURCE_INFO_NAME		= "OSPF Test";
-	protected static final String			RESOURCE_URI			= "mock://user:pass@host.net:2212/mocksubsystem";
+	protected static final String	ACTION_NAME				= "junos";
+	protected static final String	CAPABILITY_URI			= "mock://user:pass@host.net:2212/mocksubsystem";
+	protected static final String	QUEUE_CAPABILIY_TYPE	= "queue";
+	protected static final String	OSPF_CAPABILIY_TYPE		= "ospf";
+	protected static final String	OSPF_CAPABILIY_VERSION	= "10.10";
+	protected static final String	RESOURCE_TYPE			= "router";
+	protected static final String	RESOURCE_INFO_NAME		= "OSPF Test";
+	protected static final String	RESOURCE_URI			= "mock://user:pass@host.net:2212/mocksubsystem";
 
-	protected ICapability					iOSPFCapability;
-	protected IResource						routerResource;
-
-	@Inject
-	protected IResourceManager				resourceManager;
+	protected ICapability			iOSPFCapability;
+	protected IResource				routerResource;
 
 	@Inject
-	protected IProtocolManager				protocolManager;
+	protected IResourceManager		resourceManager;
 
 	@Inject
-	protected BundleContext					bundleContext;
+	protected IProtocolManager		protocolManager;
 
-    @Inject
-    @Filter("(osgi.blueprint.container.symbolicname=net.i2cat.mantychore.protocols.netconf)")
-    private BlueprintContainer				netconfService;
+	@Inject
+	protected BundleContext			bundleContext;
 
-	private static final Log					log						= LogFactory
-																			.getLog(OSPFIntegrationTest.class);
+	@Inject
+	@Filter("(osgi.blueprint.container.symbolicname=net.i2cat.mantychore.protocols.netconf)")
+	private BlueprintContainer		netconfService;
+
+	private static final Log		log						= LogFactory
+																	.getLog(OSPFIntegrationTest.class);
 
 	@Configuration
 	public static Option[] configuration() {
-		return options(karafDistributionConfiguration()
-					   .frameworkUrl(maven()
-									 .groupId("net.i2cat.mantychore")
-									 .artifactId("assembly")
-									 .type("zip")
-									 .classifier("bin")
-									 .versionAsInProject())
-					   .karafVersion("2.2.2")
-					   .name("mantychore")
-					   .unpackDirectory(new File("target/paxexam")),
-					   editConfigurationFilePut("etc/org.apache.karaf.features.cfg",
-												"featuresBoot",
-												"opennaas-router,nexus-tests-helper"),
-					   configureConsole()
-					   .ignoreLocalConsole()
-					   .ignoreRemoteShell(),
-					   keepRuntimeFolder());
+		return options(opennaasDistributionConfiguration(),
+				includeFeatures("opennaas-router, nexus-tests-helper"),
+				noConsole(),
+				keepRuntimeFolder());
 	}
 
 	@Before
@@ -105,7 +90,7 @@ public abstract class OSPFIntegrationTest
 
 	/**
 	 * Start router resource with 2 capabilities -> ospf & queue
-	 *
+	 * 
 	 * @throws ResourceException
 	 * @throws ProtocolException
 	 */
@@ -138,7 +123,7 @@ public abstract class OSPFIntegrationTest
 
 	/**
 	 * Stop and remove the router resource
-	 *
+	 * 
 	 * @throws ResourceException
 	 * @throws ProtocolException
 	 */
@@ -170,7 +155,7 @@ public abstract class OSPFIntegrationTest
 
 	/**
 	 * If not exists the protocol session manager, it's created and add the session context
-	 *
+	 * 
 	 * @param resourceId
 	 * @throws ProtocolException
 	 */
