--- manticore/tests/integration/org.opennaas.router.tests.capability.ospf/src/test/java/org/opennaas/router/tests/capability/ospf/methods/DeactivateOSPFTest.java
@@ -1,39 +1,0 @@
-/**
- *
- */
-package org.opennaas.router.tests.capability.ospf.methods;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.opennaas.core.resources.ResourceException;
-import org.opennaas.core.resources.capability.ICapability;
-import org.opennaas.core.resources.protocol.ProtocolException;
-import org.opennaas.core.resources.queue.QueueConstants;
-import org.opennaas.core.resources.queue.QueueResponse;
-import org.opennaas.router.capability.ospf.OSPFCapability;
-import org.opennaas.router.tests.capability.ospf.OSPFIntegrationTest;
-
-/**
- * @author Jordi
- */
-public class DeactivateOSPFTest extends OSPFIntegrationTest {
-
-	/**
-	 * Test to check deactivateOSPF method
-	 */
-	@Test
-	public void deactivateOSPFTest()
-		throws ResourceException, ProtocolException
-	{
-		startResource();
-
-		OSPFCapability ospfCapability = (OSPFCapability) routerResource.getCapability(getOSPFInformation(OSPF_CAPABILIY_TYPE));
-		ospfCapability.deactivateOSPF();
-
-		ICapability queueCapability = routerResource.getCapability(getOSPFInformation(QUEUE_CAPABILIY_TYPE));
-		QueueResponse queueResponse = (QueueResponse) queueCapability.sendMessage(QueueConstants.EXECUTE, null);
-		Assert.assertTrue(queueResponse.isOk());
-
-		stopResource();
-	}
-}
