--- manticore/tests/integration/org.opennaas.router.tests.capability.ospf/src/test/java/org/opennaas/router/tests/capability/ospf/methods/GetOSPFConfigurationTest.java
@@ -1,39 +1,0 @@
-/**
- *
- */
-package org.opennaas.router.tests.capability.ospf.methods;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.opennaas.core.resources.ResourceException;
-import org.opennaas.core.resources.capability.ICapability;
-import org.opennaas.core.resources.protocol.ProtocolException;
-import org.opennaas.core.resources.queue.QueueConstants;
-import org.opennaas.core.resources.queue.QueueResponse;
-import org.opennaas.router.capability.ospf.OSPFCapability;
-import org.opennaas.router.tests.capability.ospf.OSPFIntegrationTest;
-
-/**
- * @author Jordi
- */
-public class GetOSPFConfigurationTest extends OSPFIntegrationTest {
-
-	/**
-	 * Test to check getOSPFConfiguration method
-	 */
-	@Test
-	public void getOSPFConfigurationTest()
-		throws ResourceException, ProtocolException
-	{
-		startResource();
-
-		OSPFCapability ospfCapability = (OSPFCapability) routerResource.getCapability(getOSPFInformation(OSPF_CAPABILIY_TYPE));
-		ospfCapability.getOSPFConfiguration();
-
-		ICapability queueCapability = routerResource.getCapability(getOSPFInformation(QUEUE_CAPABILIY_TYPE));
-		QueueResponse queueResponse = (QueueResponse) queueCapability.sendMessage(QueueConstants.EXECUTE, null);
-		Assert.assertTrue(queueResponse.isOk());
-
-		stopResource();
-	}
-}
